requirements.txt
chromadb
sentence-transformers
rank-bm25
langchain
langchain-community
langchain-text-splitters
langchain-core
pypdf
pdfplumber
anthropic
streamlit
requests
arxiv
python-dotenv
ollama